Agregando correcciones
