--- src/app/screens/Calculator/components/Button/styles.js
@@ -1,34 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export default StyleSheet.create({
-  boxButton: {
-    width: '25%',
-    height: '20%',
-    padding: 5,
-    backgroundColor: '#4bcaf9',
-  },
-  button: {
-    flex: 1,
-    backgroundColor: '#00d564',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 20,
-  },
-  textButton: {
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-  red: {
-    backgroundColor: 'red',
-  },
-  specialButton: {
-    width: '50%',
-  },
-  del: {
-    backgroundColor: '#8a2a2a',
-  },
-  operation: {
-    backgroundColor: '#00b555',
-  },
-});
